requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
langchain-core
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
requests
numpy
